test(onboard): split provider routing helpers
